# swagger/swagger.yaml
openapi: 3.0.0

info:
  title: Personal Budget API
  description: API for managing budget envelopes and transactions.
  version: 1.0.0

servers:
  - url: http://localhost:3000

paths:

  /envelopes:
    get:
      summary: Get all envelopes
      responses:
        '200':
          description: List of all envelopes

    post:
      summary: Create a new envelope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - balance
              properties:
                title:
                  type: string
                balance:
                  type: number
      responses:
        '201':
          description: Envelope created

  /envelopes/{id}:
    get:
      summary: Get envelope by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Envelope found
        '404':
          description: Envelope not found

    put:
      summary: Update an envelope
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                balance:
                  type: number
      responses:
        '200':
          description: Envelope updated
        '404':
          description: Envelope not found

    delete:
      summary: Delete an envelope
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Envelope deleted
        '404':
          description: Envelope not found

  /envelopes/transfer:
    post:
      summary: Transfer balance between two envelopes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from
                - to
                - amount
              properties:
                from:
                  type: integer
                to:
                  type: integer
                amount:
                  type: number
      responses:
        '200':
          description: Transfer successful
        '400':
          description: Invalid request
        '404':
          description: Envelope not found

  /transactions:
    get:
      summary: Get all transactions
      responses:
        '200':
          description: List of all transactions

    post:
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - envelope_id
                - amount
                - recipient
                - description
              properties:
                envelope_id:
                  type: integer
                amount:
                  type: number
                recipient:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Transaction created
        '400':
          description: Bad request
        '404':
          description: Envelope not found

  /transactions/{id}:
    get:
      summary: Get transaction by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Transaction found
        '404':
          description: Transaction not found

    put:
      summary: Update transaction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                recipient:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Transaction updated
        '404':
          description: Transaction not found

    delete:
      summary: Delete transaction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Transaction deleted
        '404':
          description: Transaction not found